Fix packing and unpacking for Ruby 1.9.2
'Q>' and 'S>' directives are not supported before Ruby 1.9.2, more details
at https://bugs.ruby-lang.org/issues/3491

This commit fixes packing and unpacking of messages between between 126
and 65,535 bytes, replacing 'S>' with 'n'.

Note that this error wasn't pointed out by specs because 'Q>' and S>'
directives are interpreted by Ruby 1.9.2 as 'Q' and 'S'.

For an implementation of the Ruby 1.9.2 version of 'Q>' see em-websocket:
https://github.com/igrigorik/em-websocket/blob/184977ba2fa1ee448ca3dac1f630573450ef2b31/lib/em-websocket/framing07.rb#L135